requirements.txt
FLASK-WTF
flask
google-cloud-vision
google-cloud-storage
Pillow
opencv-python